Created profile view, updated shopping templates and added middleware, decorators and signals.

+++ shopping/views.py
@@ -1,17 +1,21 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import shoppingItemModel
 from .forms import shoppingItemForm
+from users.decorators import superuser_required
+from django.urls import reverse
 from django.http import HttpResponse
 
 # Create your views here. 
+@superuser_required
 def item_create(request):
-    context = {}
-    form = shoppingItemForm(request.POST or None)
-    if form.is_valid():
-        form.save()
-        # return HttpResponse('Item is successfully created')
-        return redirect('item_list')
-    context['form'] = form
+    if request.user.is_superuser:
+        context = {}
+        form = shoppingItemForm(request.POST or None)
+        if form.is_valid():
+            form.save()
+            # return HttpResponse('Item is successfully created')
+            return redirect('item_list')
+        context['form'] = form
     return render(request, "shopping/item_create.html", context)
 
 def item_list(request):
@@ -24,21 +28,25 @@ def item_detail(request, id):
     context['item'] = shoppingItemModel.objects.get(item_id = id)
     return render(request, "shopping/item_detail.html", context)
 
+@superuser_required
 def item_update(request, id):
-    context = {}
-    obj = get_object_or_404(shoppingItemModel, item_id = id)
-    form = shoppingItemForm(request.POST or None, instance=obj)
-    if form.is_valid():
-        form.save()
-        return redirect('item_detail', obj.item_id)
-    context['form']=form
+    if request.user.is_superuser:
+        context = {}
+        obj = get_object_or_404(shoppingItemModel, item_id = id)
+        form = shoppingItemForm(request.POST or None, instance=obj)
+        if form.is_valid():
+            form.save()
+            return redirect('item_detail', obj.item_id)
+        context['form']=form
     return render(request, "shopping/item_update.html", context)
-    
+
+@superuser_required 
 def item_delete(request, id):
-    context = {}
-    item = get_object_or_404(shoppingItemModel, item_id = id)
-    if request.method == "POST":
-        item.delete()
-        return redirect('item_list')
-    context['item'] = item
+    if request.user.is_superuser:
+        context = {}
+        item = get_object_or_404(shoppingItemModel, item_id = id)
+        if request.method == "POST":
+            item.delete()
+            return redirect('item_list')
+        context['item'] = item
     return render(request, "shopping/item_delete.html", context)
